refactor: 129 - move price component into different function
