+++ src/components/catalog/card/card.ts
@@ -1,9 +1,9 @@
 import { createButton, createDiv, createImg, createSpan } from '../../../utils/create-elements/create-tags';
 import { CARD } from '../../../pages/catalog/constants';
 import { BUTTONS_CONFIG } from '../../../shared/ui-config/button';
-import { CENTS_IN_DOLLAR } from '../../../shared/constants';
 import type { ProductProjection } from '@commercetools/platform-sdk';
 import { calculateDiscount } from '../../../helpers/calculate-discount';
+import { createPriceComponent } from './price';
 
 export function createProductCard(product: ProductProjection): HTMLDivElement {
   const productDescription = product.description?.en || 'Just a cool product for you!';
@@ -17,8 +17,6 @@ export function createProductCard(product: ProductProjection): HTMLDivElement {
 
   const discountPercent = calculateDiscount(productPrice, productDiscount);
 
-  const isDiscount = productDiscount > 0;
-
   const imageWrapper = createDiv({ classes: CARD.imgWrapper });
   createImg({
     classes: [...CARD.img, ...CARD.imgHover],
@@ -27,28 +25,18 @@ export function createProductCard(product: ProductProjection): HTMLDivElement {
   });
 
   if (discountPercent !== undefined) {
-    const discount = createDiv({
+    createDiv({
       text: `-${discountPercent}%`,
-      classes: [CARD.discountPercent],
+      classes: CARD.discountPercent,
+      parent: imageWrapper,
     });
-    imageWrapper.append(discount);
   }
 
   const name = createSpan({ text: productName, classes: CARD.title });
   const description = createSpan({ text: productDescription, classes: CARD.description });
 
-  const price = createSpan({
-    text: `${productPrice / CENTS_IN_DOLLAR} $`,
-    classes: [...CARD.price, ...(isDiscount ? ['line-through'] : [])],
-  });
-  const discount = isDiscount
-    ? createSpan({ text: `${productDiscount / CENTS_IN_DOLLAR} $`, classes: CARD.discount })
-    : undefined;
+  const priceChildren = createPriceComponent(productDiscount, productPrice);
 
-  const priceChildren = [price];
-  if (discount) {
-    priceChildren.push(discount);
-  }
   const priceWrapper = createDiv({ classes: CARD.priceWrapper, children: priceChildren });
 
   const basketButton = createButton(BUTTONS_CONFIG.basket);
